cmd/snap-confine: Unconditionally allow cgroup access to binder devices
binder and hwbinder are devices residing in their own filesystem.
In order to allow access to them they need to be unconditionally
allowed in the cgroup for libhybris drivers to work accordingly.

Note that binderfs is checked before binder in order to distinguish
between an Android 11+ and lower kernels since newer ones place
a compatibility symlink at the old and familiar path.